implement "auto" input for testing purposes. optimize scanForIdenticalObject using epsilon floating-point comparison upgrade cascading deletion to handle recursive nodes fix out_of_range crash by correcting ID vs index mismatch in transparent render queue

+++ Object.h
@@ -35,7 +35,7 @@ public:
 	std::vector<float>&		  getComponents()		 { return m_components; }
 	const glm::vec4&          getColor() const		 { return m_color; }
 	const std::array<int, 3>& getParentIDs() const   { return m_parentIDs; }
-	uint8_t			          getParentCount() const { return m_parentCount; }
+	int					      getParentCount() const { return m_parentCount; }
 	const std::array<int, 3>& getpCompIndex() const  { return m_pCompIndex; }
